requirements.txt
ultralytics>=8.0.0
numpy>=1.20.0
opencv-python>=4.5.0
matplotlib>=3.5.0
Pillow>=8.0.0 